Simple convolution works again
Former-commit-id: 42729cfa344b100a4252587fb85095a24081484f
Former-commit-id: 2167ad4b4e3fe1bff399432c28f0ba50a1b62981
Former-commit-id: 0fd7dbc0c10f4d69d3aa440a74044356cdb5d140

+++ benchmarks/heteromark/fir/fir.go
@@ -81,10 +81,10 @@ func (b *Benchmark) initMem() {
 	b.gHistoryData = b.driver.AllocateMemory(b.context, uint64(b.numTaps*4))
 	b.gInputData = b.driver.AllocateMemory(b.context, uint64(b.Length*4))
 	b.driver.Distribute(b.context,
-		uint64(b.gInputData), uint64(b.Length*4), b.gpus)
+		b.gInputData, uint64(b.Length*4), b.gpus)
 	b.gOutputData = b.driver.AllocateMemory(b.context, uint64(b.Length*4))
 	b.driver.Distribute(b.context,
-		uint64(b.gOutputData), uint64(b.Length*4), b.gpus)
+		b.gOutputData, uint64(b.Length*4), b.gpus)
 	b.driver.MemCopyH2D(b.context, b.gInputData, b.inputData)
 
 	for i, gpu := range b.gpus {
